added complete block/unblock user backend support

+++ Home/consumers.py
@@ -8,7 +8,10 @@ from channels.layers import get_channel_layer
 from Profile.models import User, Friends
 from Profile.tasks import create_or_update_convo_obj
 from Home.models import PostModel, PostLikes, PostComments, UserNotification, Conversations
-from Home.tasks import send_notifications, del_notifications, monitor_user_status
+from Home.tasks import (
+    send_notifications, del_notifications, 
+    monitor_user_status, share_post_to_users
+)
 import json, asyncio, pytz
 from hashlib import sha256
 from datetime import datetime, timedelta
@@ -172,23 +175,26 @@ class CentralPerkHomeConsumer(AsyncWebsocketConsumer):
     def like_post_from_wall(self, post_id):
         user = self.scope['user']
         try:
-            post = PostModel.objects.filter(post_id=post_id).prefetch_related(Prefetch('post_like_obj')).select_related('user').first()
-            if post.post_like_obj.filter(user=user).exists():
-                # Dislike post
-                post.post_like_obj.filter(user=user).delete()
-                del_notifications.delay(username=user.username, reaction="Disliked", send_to_username=post.user.username, post_id=post_id)
-                post.likes_count = F('likes_count') - 1
-                post.save()
-                post.refresh_from_db()
-            else:
-                # Like post
-                post_like_obj = PostLikes.objects.create(post_obj=post, user=user)
-                post.post_like_obj.add(post_like_obj)
-                post.likes_count = F('likes_count') + 1
-                post.save()
-                post.refresh_from_db()
-                # Notify the user whose post is being liked
-                send_notifications.delay(username=user.username, reaction="Liked", send_to_username=post.user.username, post_id=post_id)
+            post_obj = PostModel.objects.filter(post_id=post_id).prefetch_related(Prefetch('post_like_obj')).select_related('user').first()
+            post_user_block_obj = UserBlockList.objects.filter(current_user=post_obj.user).first()
+            # check if user's not being blocked by post.user
+            if not post_user_block_obj.blocked_user.filter(username=user.username).exists():
+                if post.post_like_obj.filter(user=user).exists():
+                    # Dislike post
+                    post.post_like_obj.filter(user=user).delete()
+                    del_notifications.delay(username=user.username, reaction="Disliked", send_to_username=post.user.username, post_id=post_id)
+                    post.likes_count = F('likes_count') - 1
+                    post.save()
+                    post.refresh_from_db()
+                else:
+                    # Like post
+                    post_like_obj = PostLikes.objects.create(post_obj=post, user=user)
+                    post.post_like_obj.add(post_like_obj)
+                    post.likes_count = F('likes_count') + 1
+                    post.save()
+                    post.refresh_from_db()
+                    # Notify the user whose post is being liked
+                    send_notifications.delay(username=user.username, reaction="Liked", send_to_username=post.user.username, post_id=post_id)
         except Exception as e:
             print(str(e))
 
@@ -200,14 +206,17 @@ class CentralPerkHomeConsumer(AsyncWebsocketConsumer):
                 # Cannot accept blank comments or comments with only spaces or newlines
                 return 'comment cannot be empty'
             post_obj = PostModel.objects.get_post(post_id=post_id)
-            c = PostComments.objects.create(user=user, post_obj=post_obj, comment=comment)
-            post_obj.post_comment_obj.add(c)
-            post_obj.comment_count = F('comment_count') + 1
-            post_obj.save()
-            post_obj.refresh_from_db()
-            # Notify the user whose post you commented on
-            send_notifications.delay(username=user.username, reaction='Commented', 
-            send_to_username=post_obj.user.username, post_id=post_id, comment_id=c.comment_id)
+            post_user_block_obj = UserBlockList.objects.filter(current_user=post_obj.user).first()
+            # check if user's not being blocked by post.user
+            if not post_user_block_obj.blocked_user.filter(username=user.username).exists():
+                c = PostComments.objects.create(user=user, post_obj=post_obj, comment=comment)
+                post_obj.post_comment_obj.add(c)
+                post_obj.comment_count = F('comment_count') + 1
+                post_obj.save()
+                post_obj.refresh_from_db()
+                # Notify the user whose post you commented on
+                send_notifications.delay(username=user.username, reaction='Commented', 
+                send_to_username=post_obj.user.username, post_id=post_id, comment_id=c.comment_id)
             return post_obj.comment_count
         except Exception as e:
             print(str(e))
@@ -217,12 +226,15 @@ class CentralPerkHomeConsumer(AsyncWebsocketConsumer):
         try:
             user = self.scope['user']
             post_obj = PostModel.objects.get_post(post_id=post_id)
-            if user in post_obj.saved_by.all():
-                post_obj.saved_by.remove(user)
-                return 'unsaved'
-            else:
-                post_obj.saved_by.add(user)
-                return 'saved'
+            post_user_block_obj = UserBlockList.objects.filter(current_user=post_obj.user).first()
+            # check if user's not being blocked by post.user
+            if not post_user_block_obj.blocked_user.filter(username=user.username).exists():
+                if user in post_obj.saved_by.all():
+                    post_obj.saved_by.remove(user)
+                    return 'unsaved'
+                else:
+                    post_obj.saved_by.add(user)
+                    return 'saved'
         except Exception as e:
             print(str(e))
 
@@ -240,8 +252,12 @@ class CentralPerkHomeConsumer(AsyncWebsocketConsumer):
                     # Create or update a conversation model btw request_by_user & user
                     create_or_update_convo_obj.delay(request_by.username, user.username, 'follow')
                     Friends.follow(request_by, user)
+                    # delete any existing similar notification
+                    del_notifications(reaction="Accept Follow Request", username=user.username, send_to_username=request_by.username)
                     send_notifications.delay(username=user.username, reaction="Accept Follow Request", 
                     send_to_username=request_by.username, private_request=False)
+                    # share two of user's recent posts *(if any) to request_by.user's wall
+                    share_post_to_users.delay(username=user.username, send_to=[str(request_by.username)], to_wall=True)
                 else:
                     create_or_update_convo_obj.delay(request_by.username, user.username, 'unfollow')
                 notification.delete()
